// src/service/types.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface MergeRequest {
    id:                            number;
    iid:                           number;
    project_id:                    number;
    title:                         string;
    description:                   string;
    state:                         string;
    created_at:                    Date;
    updated_at:                    Date;
    merged_by:                     null;
    merge_user:                    null;
    merged_at:                     null;
    closed_by:                     null;
    closed_at:                     null;
    target_branch:                 string;
    source_branch:                 string;
    user_notes_count:              number;
    upvotes:                       number;
    downvotes:                     number;
    author:                        Author;
    assignees:                     any[];
    assignee:                      null;
    reviewers:                     any[];
    source_project_id:             number;
    target_project_id:             number;
    labels:                        any[];
    draft:                         boolean;
    work_in_progress:              boolean;
    milestone:                     null;
    merge_when_pipeline_succeeds:  boolean;
    merge_status:                  string;
    detailed_merge_status:         string;
    sha:                           string;
    merge_commit_sha:              null;
    squash_commit_sha:             null;
    discussion_locked:             null;
    should_remove_source_branch:   null;
    force_remove_source_branch:    boolean;
    reference:                     string;
    references:                    References;
    web_url:                       string;
    time_stats:                    TimeStats;
    squash:                        boolean;
    squash_on_merge:               boolean;
    task_completion_status:        TaskCompletionStatus;
    has_conflicts:                 boolean;
    blocking_discussions_resolved: boolean;
    approvals_before_merge:        null;
}

export interface Author {
    id:         number;
    username:   string;
    name:       string;
    state:      string;
    avatar_url: string;
    web_url:    string;
}

export interface References {
    short:    string;
    relative: string;
    full:     string;
}

export interface TaskCompletionStatus {
    count:           number;
    completed_count: number;
}

export interface TimeStats {
    time_estimate:          number;
    total_time_spent:       number;
    human_time_estimate:    null;
    human_total_time_spent: null;
}